PIX-20 Refactor
