+++ packages/client/src/features/ColumnHeader/index.ts
@@ -1,0 +1,1 @@
+export { ColumnHeader } from './ColumnHeader';
