+++ packages/client/src/pages/LeaderBoardPage/LeaderBoardPage.tsx
@@ -1,6 +1,6 @@
 import { DataGrid, GridRowsProp, GridColDef } from '@mui/x-data-grid';
-import { ColumnHeader } from './components/ColumnHeader';
-import type { ILeader } from '@/shared/types';
+import { ColumnHeader } from '@/features/ColumnHeader';
+import type { ILeader } from '@/shared/types/leaderboard.interface';
 
 const mockRows: GridRowsProp<ILeader> = [
   { id: '1', firstName: 'Luke', lastName: 'Skywalker', score: 1350 },
