+++ packages/client/src/shared/types/auth.interface.ts
@@ -1,10 +1,3 @@
-export type ILeader = {
-  id: string;
-  firstName: string;
-  lastName: string;
-  score: number;
-};
-
 export type IUser = {
   id: string;
   first_name: string;
