+++ packages/client/src/shared/types/leaderboard.interface.ts
@@ -1,0 +1,6 @@
+export type ILeader = {
+  id: string;
+  firstName: string;
+  lastName: string;
+  score: number;
+};
